requirements.txt
coqui-tts
torch
torchaudio
pandas
numpy
librosa
soundfile
scikit-learn
matplotlib
tensorboard
pydub